refactor: reorganize ServiceMarketWS structure and enhance type definitions
- Deleted the old ServiceMarketWS implementation and created a new file structure for better organization.
- Introduced new types for price and transaction data, improving clarity and maintainability.
- Updated the export structure to streamline imports and enhance code readability.
