+++ packages/kit-bg/src/services/ServiceMarketWS/types/index.ts
@@ -1,0 +1,1 @@
+export * from './priceData';
