+++ packages/kit-bg/src/services/ServiceMarketWS/types/priceData.ts
